Appのstateや関数をMemoListに移した
・AppからMemoListに対して多くのpropsを渡していたが、それらがAppに定義されている必要がなかったため、MemoListに移した。
